authentication nonces

+++ inc/authentication.php
@@ -99,12 +99,16 @@ class blippress_authentication {
 		$opts = get_option( $this->option() );
 
 		if ( isset( $_POST['revoke-permission'] ) and 'go' == $_POST['revoke-permission'] ) {
+
+			check_admin_referer( blippress_prefix() . 'revoke-permission-nonce', blippress_prefix() . 'revoke-permission-hash' );
 			delete_option( $this->option() );
 			$blippress_cache->clear();
+
 		}
 
 		if ( isset( $_POST['request-permission'] ) and 'go' == $_POST['request-permission'] ) {
 
+			check_admin_referer( blippress_prefix() . 'permission-nonce', blippress_prefix() . 'permission-hash' );
 			$blip = new blipWP( $blippress->key );
 			$blip->get_temp_token( $blippress->permissions_id, $this->page_url() );
 
@@ -154,6 +158,7 @@ class blippress_authentication {
 				<p>You need to grant permission for your website to access your Blipfoto account.</p>
 				<form method="post">
 					<p>
+						<?php wp_nonce_field( blippress_prefix() . 'permission-nonce', blippress_prefix() . 'permission-hash' ); ?>
 						<input type="hidden" name="request-permission" value="go">
 						<input class="button-primary" name="submit" type="submit" value="Grant permission">
 					</p>
@@ -161,9 +166,12 @@ class blippress_authentication {
 
 				<h4>What does this mean?</h4>
 
-				<p>BlipPress needs your permission in order to access your Blipfoto account, and to prevent anyone accessing your account without your permission. When you click the button above, you will be redirected tothe 'apps' page on blipfoto.com where you can verify that you are happy to allow it. You only need to do this once.</p>
+				<p>BlipPress needs your permission in order to access your Blipfoto account, and to prevent anyone accessing your account without your permission. When you click the button above, you will be redirected to the 'apps' page on blipfoto.com where you can verify that you are happy to allow it. You only need to do this once.</p>
 				<p>You will then be able to display blips here on your website and create entries on your Blipfoto journal from the post edit screen.</p>
-				<p>BlipPress does not have the ability to create entries by itself, nor does it allow anyone else to, only you can manually do this. The creators of BlipPress have no way of accessing any of your account credentials or other personal details (the necessary access token and key are only stored here in your website's database).</p>
+				<h4>Privacy</h4>
+				<p>BlipPress does not have the ability to create entries by itself, nor does it allow anyone else to, only you can manually do this.</p>
+				<p>The creators of BlipPress have no way of accessing any of your account credentials or other personal details (the necessary access token and key are only stored here in your website's database).</p>
+				<h4>Questions?</h4>
 				<p>If you have any questions about this, please use the support forums at the <a href="<?php echo blippress_plugin_page(); ?>">plugin page</a>.</p>
 
 			<?php } else { ?>
@@ -180,6 +188,7 @@ class blippress_authentication {
 				<p><strong>Important!</strong> Revoking permission will prevent any blips being displayed on your website and you will not be able to post to Blipfoto until you grant permission again. If you change to a different account, some blips displayed on your site may change.</p>
 				<form method="post">
 					<p>
+						<?php wp_nonce_field( blippress_prefix() . 'revoke-permission-nonce', blippress_prefix() . 'revoke-permission-hash' ); ?>
 						<input type="hidden" name="revoke-permission" value="go">
 						<input class="button-primary" name="submit" type="submit" value="Revoke permission" onClick="return confirm('Are you sure you want to revoke permission?')">
 					</p>
